# Enable optimisation in Cargo.toml of b3sum