--- a/.github/workflows/build_b3sum.py
+++ b/.github/workflows/build_b3sum.py
@@ -9,7 +9,7 @@ import sys
 ROOT = Path(__file__).parent.parent.parent
 RUST_TARGET = sys.argv[1]
 
-subprocess.run(["cargo", "build", "--target", sys.argv[1], "--release"],
+subprocess.run(["cargo", "build", "--target", sys.argv[1], "--release", "--config", "strip=true"],
                cwd=ROOT / "b3sum")
 
 if platform.system() == "Windows":