--- a/b3sum/Cargo.toml
+++ b/b3sum/Cargo.toml
@@ -25,3 +25,8 @@ wild = "2.0.3"
 [dev-dependencies]
 duct = "0.13.3"
 tempfile = "3.1.0"
+
+[profile.release]
+strip = true
+lto = true
+opt-level = 3