@oconnor663 2022-09-12
Can we compare this to what some popular Rust binary crates are doing? In particular, it looks like `ripgrep` [doesn't turn on these options](https://github.com/BurntSushi/ripgrep/blob/a66315d232b49e623268de801af19460108ba514/Cargo.toml#L79-L80). Maybe the idea is that individual packagers can turn on the options they prefer using environment variables? @burntsushi I'd be curious to get your advice about this. I don't see much when I search for "rust lto best practices" or similar.