@oconnor663 2022-09-12
Regarding performance testing, I don't expect any of these flags to have much impact in the default case either way. The most performance-critical parts are written in assembly, which won't be affected by these flags. However, it might be interesting to test the pure/portable Rust implementation, by measuring with `--features pure`.