@Yusuto 2022-09-14
From what I could see, compile times are largely unaffected here by LTO, but as you said so is performance. I think the things that I found more peculiar were stripping and optimization, especially the former, since it makes the file four times smaller, which I think is nice. For the other two I think I agree with »If it doesn't matter, why bother?«.