@Yusuto 2022-09-16
Should we maybe just remove LTO and Compiler Optimizations and only enable stripping? That part just seems like a free 75% reduction in file size, right?