@oconnor663 2022-10-01
How about we enable stripping in our internal CI build (the binaries that get hosted here on GitHub and attached to each release tag) but keep the defaults in `Cargo.toml`?