@Yusuto 2022-10-02
I think that actually sounds like a very good idea, Yes!
That can probably be easily automated, right? There is probably some option for enabling stripping on the command-line, something like `cargo build --release --profile strip=true` (that doesn't actually work, sadly) in the CI?