@oconnor663 2022-10-02
Looks like we might want to copy something like what `ripgrep` does here: https://github.com/BurntSushi/ripgrep/blob/4386b8e805e273a9795ad4e256c455c74407c949/.github/workflows/release.yml#L135-L146

And put it in between these two steps here: https://github.com/BLAKE3-team/BLAKE3/blob/7e15c5314ea0ca34c52fd39583b1aa7c07eac1ae/.github/workflows/tag.yml#L38-L45