@Yusuto 2022-10-02
Hmm, couldn't we just let Cargo take care of stripping, now that the feature is stabilized? I finally got to look up how, it seems to be `cargo build --release --config strip=true`, and that would need to be changed in [`build_b3sum.py`](https://github.com/BLAKE3-team/BLAKE3/blob/7e15c5314ea0ca34c52fd39583b1aa7c07eac1ae/.github/workflows/build_b3sum.py#L12)